// src/app/types/Vehicle.ts
export class Vehicle {
  id: number;
  model: string;
  make: string;
  yearOfMake: number;
  engineNo: string;
  chassisNo: string;
  plateNo: string;
  cc: string;
  fuelType: number;
  disPerLit: number;
  carryingCapacity: string;

  lastMilege: number;
  insExpDate: Date;
  insRenewalDate: Date;
  policyNo: string;
  lmGenService: number;
  lmTyresChange: number;
  lmVehicleBody: number;
  description: string;
  purposeOfUsage: string;
  status: string;
  owner: string;
  custodian: string;
  previousServiceDate:Date;
  vehicleTypeString:string;
  vehicleType:number;
  createdBy:string;
  createdDate:Date;
  isStandBy:string;
}